extract stepper and styles into UI
